+++ features/nyxus-tool/CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.1.8-dev1] - 2024-11-15
+- fix a bug for no objects in label images.
+- updated bfio base-container image.
+
 ## [0.1.8-dev0] - 2024-09-26
 - Renamed nyxus-plugin to nyxus-tool.
 - updated nyxus package and bfio base-container image.